requirements.txt
# Core RAG Dependencies (Updated for Python 3.13 compatibility)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# Vector Database
chromadb>=0.5.0

# Embeddings
sentence-transformers>=2.3.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM Integration
openai>=1.10.0
tiktoken>=0.5.0

# UI Framework
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0

# Basic dependencies
numpy>=1.26.0
pandas>=2.0.0

